requirements.txt
fastapi>=0.115.8
google-genai>=1.5.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
python-multipart>=0.0.20
uvicorn[standard]>=0.34.0
